refactor(spec): adjusted types based on return values
